Extract custom errors into module & delete old code
