+++ lib/highline/custom_errors.rb
@@ -1,0 +1,20 @@
+module CustomErrors
+  # Internal HighLine errors.
+  class QuestionError < StandardError
+  end
+
+  class NotValidQuestionError < QuestionError
+  end
+
+  class NotInRangeQuestionError < QuestionError
+  end
+
+  class NoConfirmationQuestionError < QuestionError
+  end
+
+  class NoAutoCompleteMatch < StandardError
+  end
+
+  class NoAutoCompleteMatch < StandardError
+  end
+end
